Fixed same assert error as before in 03, 04 and 05.
The MySQL dataset timestamps (in the last_update columns) differ from
the timestamps in the PostgreSQL database (one takes GMT offset into
account). So one shouldn't test for a match on the timestamp, only
on the date part.  Changed it to do a regex match just on the date part.