Updated ant build.xml to work with new dir structure and H2.
Has a series of targets for "completed koans" and for compiling
and running the RunH2 java class.  These all depend on having
the jars in (or linked in) the lib directory of this project -
they do not depend on maven in any way.

Also updated koans16 and 17 to fix minor path issues - they now
pass with MySQL and PostgreSQL from the ant command line.  H2
koans 16 and 17 still not present/passing.